refactor - move showdown extension
